// src/enums/mastercode/relation-status.enum.ts
/**
 * 친구과의 관계를 정의하는 열거형
 *
 * @author joohongpark
 */
const enum RelationStatus {
  FRST10 = 'FRST10', // 친구 수락
  FRST20 = 'FRST20', // 친구 거절
  FRST30 = 'FRST30', // 친구 차단
  FRST40 = 'FRST40', // 지인
  FRST50 = 'FRST50', // 절교
}
export default RelationStatus;